refactor(core): split shapes.rs and get every dir to <=5 files
Bring funput-core to the repo standard (<=150 LOC/file, <=5 files/dir). Pure
internal reorganization of the module tree — no public API or behavior change.

- unicode/shapes.rs (202 LOC): extract the inline tests to unicode/shapes/tests.rs
- composition/transform, composition/uo_horn: move tests.rs into a tests/ subdir (6->5)
- composition/intent (7): group the three per-variant resolvers (circumflex, stroke,
  w) under intent/kinds/, leaving the shared candidate/target machinery at the root (->4)
- input_method/telex (6): group the four modifier classifiers (circumflex, stroke,
  tone, w) under telex/modifiers/, leaving advanced at the root (->2 + subdirs)
- validation (6): merge syllable.rs into the existing syllable/ dir as syllable/mod.rs (->5)

The crate's public surface (lib.rs) is unchanged, so funput-engine and every other
consumer compile untouched. Git tracks the moves as renames.
