+++ crates/funput-core/src/composition/intent/kinds/mod.rs
@@ -1,0 +1,6 @@
+//! The per-variant modifier-intent resolvers dispatched by [`super::ModifierIntent`]:
+//! a free-position circumflex, the `đ` stroke, and a deferred Telex `w`.
+
+pub(crate) mod circumflex;
+pub(crate) mod stroke;
+pub(crate) mod w;
