+++ crates/funput-core/src/input_method/telex/mod.rs
@@ -2,14 +2,11 @@
 //! common dispatch remains small and its priority is explicit.
 
 mod advanced;
-mod circumflex;
-mod stroke;
-mod tone;
-mod w;
+mod modifiers;
 
 use crate::input_method::{AdvancedAction, KeyAction};
 
-pub use tone::tone_from_key;
+pub use modifiers::tone::tone_from_key;
 
 pub fn classify_advanced_key(buffer: &str, key: char) -> AdvancedAction {
     advanced::classify(buffer, key)
@@ -20,25 +17,25 @@ pub fn classify_key(buffer: &str, key: char) -> KeyAction {
     if crate::composition::intent::has_pending(buffer) {
         return KeyAction::DeferredW;
     }
-    if let Some(action) = stroke::classify(buffer, key) {
+    if let Some(action) = modifiers::stroke::classify(buffer, key) {
         return action;
     }
-    if let Some(action) = circumflex::classify(buffer, key) {
+    if let Some(action) = modifiers::circumflex::classify(buffer, key) {
         return action;
     }
     if key.eq_ignore_ascii_case(&'w')
-        && let Some(action) = w::classify(buffer)
+        && let Some(action) = modifiers::w::classify(buffer)
     {
         return action;
     }
     if key.eq_ignore_ascii_case(&'z') {
         return KeyAction::RemoveTone;
     }
-    tone::classify(buffer, key).unwrap_or(KeyAction::Normal)
+    modifiers::tone::classify(buffer, key).unwrap_or(KeyAction::Normal)
 }
 
 pub(crate) fn classify_w(buffer: &str) -> Option<KeyAction> {
-    w::classify(buffer)
+    modifiers::w::classify(buffer)
 }
 
 pub(super) fn last_char(buffer: &str) -> Option<char> {
