+++ crates/funput-core/src/input_method/telex/modifiers/w.rs
@@ -3,7 +3,7 @@ use crate::input_method::KeyAction;
 use crate::unicode::marks::{is_vowel, tone_on_vowel, vowel_stem};
 use crate::unicode::shapes::{VowelShape, shape_on_vowel, shape_target_index};
 
-use super::last_char;
+use super::super::last_char;
 
 fn plain(ch: char, stem: char) -> bool {
     vowel_stem(ch).is_some_and(|value| value.eq_ignore_ascii_case(&stem))
@@ -23,7 +23,7 @@ fn ends_with_ua_target(buffer: &str) -> bool {
     is_u && !matches!(rev.next(), Some('q' | 'Q'))
 }
 
-pub(super) fn classify(buffer: &str) -> Option<KeyAction> {
+pub(crate) fn classify(buffer: &str) -> Option<KeyAction> {
     if uo_pair_in_vowel_cluster(buffer).is_some() {
         return horn();
     }
